Implement listing 62-4
Add tty_functions module containing utility functions for getting
the current settings for a terminal associated with a file
descriptor and putting the associated terminal into cbreak and raw
modes.

Use these to implement listing 62-4 which puts the terminal into
raw or cbreak mode depending on whether a command-line argument is
provided. Signals are not generated in raw mode so only quitting
via a 'q' input will cause the program to exit normally (SIGTERM
is also handled and restores the previous settings).

Signals are generated in cbreak mode so these are handled normally
and restore the previous user settings before exit. The SIGCONT and
SIGTSTP signals are also handled and save/restore the user settings
around the suspend/resume signals.