Implement listings 64-1, 64-2 and 64-3
Implement listing 64-1 which contains a utility function for creating
a master/slave pseudo-terminal pair.

Implement listing 64-2 which creates a pty, forks a new process and
sets the slave terminal as the controlling terminal for the new
child.

Implement listing 64-3 which contains a basic implementation of the
script program. This saves the current terminal settings, creates
a new pty before forking a new shell as a child process. The
parent process then continually reads from the output on the master
pty (connected to stdout in the new shell) and its own stdin
(connected to stdin). Input on stdin is copied to the pty while any
output from the child shell is copied to the session file and stdout.