Implement listing 44-4
Implement listing 44-4 which creates two child process which share
a pipe. The first child executes 'ls' with stdout bound to the
write end of the pipe, while the second child executes 'wc -l'
with stdin bound to the read end.

Move create_pipe function into util namespace so it can be used in
both the pipe_sync and pipe_ls_wc programs.