Implement listings 44-6 and 44-7
Implement listings 44-6 and 44-7 which implement a server and client
respectively which communicate over named pipes. The server creates
a named pipe to which clients send requests. Each client also create
a named pipe with a location based on the client PID. This PID is
included in the request which the server uses to construct the client
channel.