Implement listings 46-8 and 46-9
Implement listing 46-8 which defines a file server which
listens on a known message queue address for file requests. Files
are sent in a sequence of responses until the file is transmitted.

Implement listing 46-9 which implements a client for the file server.
The file to receive is specified on the command line and a request
message is sent to the server containing the file to receive and the
id of the client message queue. The server sends file response (or
error) messages to the client on its queue.