Implement listings 57-5, 57-6 and 57-7
Implement listing 57-6 containing a server which listens on a unix
socket for incoming datagram requests. These are expected to contain
strings which are converted to uppercase and returned to the client.

Implement listing 57-7 which sends a collection of strings specified
on the command line to the server and displays the responses.